+++ packages/world/src/Types.sol
@@ -110,8 +110,8 @@ struct TransferCommonContext {
   bytes32 playerEntityId;
   bytes32 chestEntityId;
   VoxelCoord chestCoord;
-  uint8 chestObjectTypeId;
-  uint8 dstObjectTypeId;
+  uint16 chestObjectTypeId;
+  uint16 dstObjectTypeId;
   ChipData checkChipData;
   bool isDeposit;
 }
@@ -119,5 +119,5 @@ struct TransferCommonContext {
 struct PipeTransferCommonContext {
   VoxelCoord targetCoord;
   ChipData targetChipData;
-  uint8 targetObjectTypeId;
+  uint16 targetObjectTypeId;
 }
