+++ packages/world/src/codegen/tables/PlayerStatus.sol
@@ -68,7 +68,7 @@ library PlayerStatus {
     _keyTuple[0] = entityId;
 
     bytes32 _blob = StoreSwitch.getStaticField(_tableId, _keyTuple, 0, _fieldLayout);
-    return (_toBool(uint8(bytes1(_blob))));
+    return (_toBool(uint16(bytes1(_blob))));
   }
 
   /**
@@ -79,7 +79,7 @@ library PlayerStatus {
     _keyTuple[0] = entityId;
 
     bytes32 _blob = StoreCore.getStaticField(_tableId, _keyTuple, 0, _fieldLayout);
-    return (_toBool(uint8(bytes1(_blob))));
+    return (_toBool(uint16(bytes1(_blob))));
   }
 
   /**
@@ -90,7 +90,7 @@ library PlayerStatus {
     _keyTuple[0] = entityId;
 
     bytes32 _blob = StoreSwitch.getStaticField(_tableId, _keyTuple, 0, _fieldLayout);
-    return (_toBool(uint8(bytes1(_blob))));
+    return (_toBool(uint16(bytes1(_blob))));
   }
 
   /**
@@ -101,7 +101,7 @@ library PlayerStatus {
     _keyTuple[0] = entityId;
 
     bytes32 _blob = StoreCore.getStaticField(_tableId, _keyTuple, 0, _fieldLayout);
-    return (_toBool(uint8(bytes1(_blob))));
+    return (_toBool(uint16(bytes1(_blob))));
   }
 
   /**
@@ -200,11 +200,11 @@ library PlayerStatus {
 
 /**
  * @notice Cast a value to a bool.
- * @dev Boolean values are encoded as uint8 (1 = true, 0 = false), but Solidity doesn't allow casting between uint8 and bool.
- * @param value The uint8 value to convert.
+ * @dev Boolean values are encoded as uint16 (1 = true, 0 = false), but Solidity doesn't allow casting between uint16 and bool.
+ * @param value The uint16 value to convert.
  * @return result The boolean value.
  */
-function _toBool(uint8 value) pure returns (bool result) {
+function _toBool(uint16 value) pure returns (bool result) {
   assembly {
     result := value
   }
